requirements.txt
-i https://pypi.org/simple
absl-py==0.12.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.1; python_version ~= '3.5'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cloudml-hypertune==0.1.0.dev6
cycler==0.10.0
decorator==4.4.2
flatbuffers==1.12
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core==1.26.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.4.3; python_version >= '3.6'
google-auth==1.28.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-core==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-storage==1.36.2
google-crc32c==1.1.2; python_version >= '3.5'
google-pasta==0.2.0
google-resumable-media==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.53.0; python_version >= '3.6'
grpcio==1.32.0
h5py==2.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
imageio==2.9.0
imgaug==0.4.0
ipython-genutils==0.2.0
ipython==7.12
jedi==0.18.0; python_version >= '3.6'
joblib==1.0.1; python_version >= '3.6'
keras-preprocessing==1.1.2
kiwisolver==1.3.1; python_version >= '3.6'
lxml==4.6.2
markdown==3.3.4; python_version >= '3.6'
matplotlib==3.3.4
more-itertools==8.7.0; python_version >= '3.5'
mung==1.1
muscima==0.10.0
music21==6.7.1
networkx==2.5; python_version >= '3.6'
numpy==1.19.5
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
omrdatasettools==1.3.1
opencv-python==4.5.1.48; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.1; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==8.1.2; python_version >= '3.6'
prompt-toolkit==3.0.17; python_full_version >= '3.6.1'
protobuf==3.15.6
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.8.1; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pywavelets==1.1.1; python_version >= '3.5'
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.7.2; python_version >= '3.6'
scikit-image==0.17.2
scikit-learn==0.24.1; python_version >= '3.6'
scipy==1.6.1; python_version >= '3.7'
shapely==1.7.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==2.4.0
tensorflow==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0; python_version >= '3.5'
tifffile==2021.3.17; python_version >= '3.7'
tqdm==4.59.0
traitlets==5.0.5; python_version >= '3.7'
typing-extensions==3.7.4.3
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
webcolors==1.11.1; python_version >= '3.5'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.36.2; python_version >= '3'
wrapt==1.12.1
